tests/wpt/web-platform-tests/websockets/Send-before-open.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Send WebSocket - Send before connection in established</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.js?pipe=sub"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">
        test(function () {
            var wsocket = CreateWebSocket(false, false, false);
            assert_throws("INVALID_STATE_ERR", function () { wsocket.send("Message to send") });
        }, "W3C WebSocket API - Send data on a WebSocket before connection is opened - INVALID_STATE_ERR is returned")
    </script>
</body>
</html>
